+++ examples/building-a-data-app-with-cube-workshop/dashboard-app/src/utils/isEmpty.js
@@ -1,0 +1,1 @@
+export const isEmpty = (obj) => Object.keys(obj).length === 0;